Refactor daemon scheduler into module
Move scheduling, retry, and renewal helpers into daemon.rs to separate
runtime orchestration from CLI setup. Update main tests to reference the
daemon module utilities.

Closes #34
